feat: verify proof using test verifier
