+++ deploy/002_deploy_verifier_test.ts
@@ -1,6 +1,6 @@
 import { HardhatRuntimeEnvironment } from "hardhat/types";
 import { DeployFunction } from "hardhat-deploy/types";
-import { getGroups, isDevNet } from "../common";
+import { isDevNet } from "../common";
 
 const func: DeployFunction = async function (hre: HardhatRuntimeEnvironment) {
   const { deployments, getUnnamedAccounts } = hre;
@@ -8,22 +8,13 @@ const func: DeployFunction = async function (hre: HardhatRuntimeEnvironment) {
 
   const [deployer] = await getUnnamedAccounts();
 
-  const interepTest = await deploy("InterepTest", {
+  await deploy("VerifierTest", {
     from: deployer,
     log: true,
-    args: [],
   });
-
-  const contract = await hre.ethers.getContractAt(
-    "InterepTest",
-    interepTest.address
-  );
-  const groups = getGroups();
-  const groupInsertionTx = await contract.updateGroups(groups);
-  await groupInsertionTx.wait();
 };
 export default func;
-func.tags = ["InterepTest"];
+func.tags = ["VerifierTest"];
 // skip when running on mainnet
 func.skip = async (hre: HardhatRuntimeEnvironment) => {
   if (isDevNet(hre.network.name)) {
